moving DEFAULT_SUBSET to util/definitions.py
