+++ src/datumaro/__init__.py
@@ -41,7 +41,6 @@ from .components.dataset import (
 from .components.dataset_storage import DatasetPatch
 from .components.dataset_item_storage import ItemStatus
 from .components.dataset_base import (
-    DEFAULT_SUBSET_NAME,
     CategoriesInfo,
     DatasetBase,
     DatasetItem,
@@ -52,6 +51,7 @@ from .components.dataset_base import (
     SubsetBase,
     Transform,
 )
+from .util.definitions import DEFAULT_SUBSET_NAME
 from .components.environment import Environment, PluginRegistry
 from .components.exporter import Exporter, ExportErrorPolicy, FailingExportErrorPolicy
 from .components.hl_ops import (  # pylint: disable=redefined-builtin
