+++ src/datumaro/plugins/data_formats/datumaro/exporter.py
@@ -33,7 +33,8 @@ from datumaro.components.annotation import (
     _Shape,
 )
 from datumaro.components.dataset_item_storage import ItemStatus
-from datumaro.components.dataset_base import DEFAULT_SUBSET_NAME, CategoriesInfo, DatasetItem
+from datumaro.components.dataset_base import CategoriesInfo, DatasetItem
+from datumaro.util.definitions import DEFAULT_SUBSET_NAME
 from datumaro.components.exporter import Exporter
 from datumaro.components.media import Image, MediaElement, PointCloud
 from datumaro.util import cast, dump_json_file
