+++ src/datumaro/plugins/data_formats/mapillary_vistas/importer.py
@@ -5,7 +5,8 @@ import glob
 import logging as log
 import os.path as osp
 
-from datumaro.components.dataset_base import DEFAULT_SUBSET_NAME, Importer
+from datumaro.components.dataset_base import Importer
+from datumaro.util.definitions import DEFAULT_SUBSET_NAME
 
 from .base import MapillaryVistasInstancesBase, MapillaryVistasPanopticBase
 from .format import MapillaryVistasPath, MapillaryVistasTask
