+++ src/datumaro/plugins/data_formats/yolo/exporter.py
@@ -26,7 +26,8 @@ from datumaro.components.annotation import (
 )
 from datumaro.components.dataset_storage import DatasetPatch
 from datumaro.components.dataset_item_storage import ItemStatus
-from datumaro.components.dataset_base import DEFAULT_SUBSET_NAME, DatasetItem, IDataset
+from datumaro.components.dataset_base import DatasetItem, IDataset
+from datumaro.util.definitions import DEFAULT_SUBSET_NAME
 from datumaro.components.errors import DatasetExportError, MediaTypeError
 from datumaro.components.exporter import Exporter
 from datumaro.components.media import Image
