+++ src/datumaro/plugins/splitter.py
@@ -11,7 +11,8 @@ import numpy as np
 
 from datumaro.components.annotation import AnnotationType
 from datumaro.components.cli_plugin import CliPlugin
-from datumaro.components.dataset_base import DEFAULT_SUBSET_NAME, Transform
+from datumaro.components.dataset_base import Transform
+from datumaro.util.definitions import DEFAULT_SUBSET_NAME
 from datumaro.util import cast
 
 NEAR_ZERO = 1e-7
