+++ src/datumaro/plugins/transforms.py
@@ -37,12 +37,12 @@ from datumaro.components.annotation import (
 )
 from datumaro.components.cli_plugin import CliPlugin
 from datumaro.components.dataset_base import (
-    DEFAULT_SUBSET_NAME,
     DatasetItem,
     IDataset,
     ItemTransform,
     Transform,
 )
+from datumaro.util.definitions import DEFAULT_SUBSET_NAME
 from datumaro.components.errors import DatumaroError
 from datumaro.components.media import Image
 from datumaro.util import NOTSET, filter_dict, parse_str_enum_value, take_by
