+++ src/datumaro/util/definitions.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+DEFAULT_SUBSET_NAME = "default"
