+++ tests/unit/test_compare.py
@@ -13,7 +13,8 @@ from datumaro.components.annotation import (
     PointsCategories,
     Skeleton,
 )
-from datumaro.components.dataset_base import DEFAULT_SUBSET_NAME, DatasetItem
+from datumaro.components.dataset_base import DatasetItem
+from datumaro.util.definitions import DEFAULT_SUBSET_NAME
 from datumaro.components.media import Image
 from datumaro.components.operations import DistanceComparator, ExactComparator
 from datumaro.components.project import Dataset
